osx: 64bit: fix alarm warnings

+++ Source/OS9exec_core/alarms.h
@@ -55,14 +55,14 @@ void init_alarms(void);
 void A_Insert(alarm_typ *aa);
 void A_Remove(alarm_typ *aa);
 os9err
-     A_Make(ushort pid, ulong *aId, ushort aCode, ulong aTicks, Boolean cyclic);
+     A_Make(ushort pid, uint32_t *aId, ushort aCode, uint32_t aTicks, Boolean cyclic);
 void A_Kill(ushort pid);
 
 os9err Alarm(ushort pid,
-             ulong *aId,
+             uint32_t *aId,
              short  aFunc,
              ushort aCode,
-             ulong  aTime,
-             ulong  aDate);
+             uint32_t aTime,
+             uint32_t aDate);
 
 /* eof */
